Split out the backgrounds and base screen classes from MiscScreens.h
